Port Haskell version to Rust
This is a port of the Haskell program to Rust.

https://github.com/tkoster/gameoflife-hs

Techniques and algorithms are almost identical, namely:
- naive neighbour counting algorithm (no good cache locality)
- render in software to a texture
- use SDL for window management, input events and final blit

The main loop of the Rust version is written in a zero-allocations style
however (the only allocations occur at startup). This means the Rust
version currently has about 2x better performance than the Haskell
version so there was room to increase the simulation rate and frame rate
to 120 Hz (up from 60 Hz of the Haskell version).